refactor: move SpinalSimSpiXdrMaster
Also moves ddr package files to ddr folder
