+++ lib/src/main/scala/spinal/lib/com/spi/ddr/Apb3SpiXdrMasterCtrl.scala
