+++ lib/src/main/scala/spinal/lib/com/spi/ddr/BmbSpiXdrMasterCtrl.scala
